content-visibility: Force range base/extent when computing visual selection.
Some of the code that does visual selection ends up updating style and
layout for node. This means that it will temporarily unlock c-v nodes
and may cause a state rewind from layout clean to visual update pending.

That's not an operation we support, verified by DCHECKs. So, instead
we should unlock any c-v nodes prior to getting to layout clean.

R=chrishtr@chromium.org, yosin@chromium.org
Bug: 1237533
Change-Id: Ib30036c4536bea3da2ae4fa54c19ad5684829597